requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
beautifulsoup4
pytest
jinja2
google-api-python-client
